+++ tests/model/network_model_test.py
@@ -1,8 +1,7 @@
+import os
 import pathlib
 import unittest
 from unittest.mock import patch
-
-from PySide6.QtCore import QSettings, QCoreApplication
 
 from src import settings
 from src.model.network_model import NetworkModel
@@ -13,16 +12,16 @@ class NetworkModelTest(unittest.TestCase):
 
     def setUp(self):
         """Metodo che viene chiamato prima di ogni metodo"""
-        self.env_settings = QSettings()
-        QCoreApplication.setOrganizationName("MercurySeven")
-        QCoreApplication.setApplicationName("SSD")
-        self.path = str(pathlib.Path().absolute()) + "/tests"
+        self.path = os.path.join(str(pathlib.Path().absolute()), "tests")
         self.path = r'%s' % self.path
         pathlib.Path(self.path).mkdir(parents=True, exist_ok=True)
-        self.env_settings.setValue("sync_path", self.path)
-        settings.file_name = "tests/config.ini"
+        settings.file_name = os.path.join(self.path, "config.ini")
         settings.check_file()
         self.model_test = NetworkModel()
+
+    def tearDown(self):
+        """Metodo che viene chiamato dopo ogni metodo"""
+        os.remove(settings.file_name)
 
     @patch('src.network.cookie_session.CookieSession._login', return_value=False)
     def test_login_value_exception(self, mocked_function):
