+++ tests/model/settings_model_test.py
@@ -2,8 +2,6 @@ import pathlib
 import unittest
 import os
 
-from PySide6.QtCore import QSettings, QCoreApplication
-
 from src.model.settings_model import SettingsModel
 from src.network.policy import Policy
 from src import settings
@@ -13,16 +11,13 @@ class TestSettings(unittest.TestCase):
 
     def setUp(self):
         """Metodo che viene chiamato prima di ogni metodo"""
-        self.env_settings = QSettings()
-        QCoreApplication.setOrganizationName("MercurySeven")
-        QCoreApplication.setApplicationName("SSD")
-        self.sett_model = SettingsModel()
-        self.path = str(pathlib.Path().absolute()) + "/tests"
+        self.path = os.path.join(str(pathlib.Path().absolute()), "tests")
         self.path = r'%s' % self.path
         pathlib.Path(self.path).mkdir(parents=True, exist_ok=True)
-        self.env_settings.setValue("sync_path", self.path)
-        settings.file_name = "tests/config.ini"
+        settings.file_name = os.path.join(self.path, "config.ini")
         settings.check_file()
+
+        self.sett_model = SettingsModel()
 
     def tearDown(self):
         """Metodo che viene chiamato dopo ogni metodo"""
