+++ tests/view/login_screen_test.py
@@ -1,7 +1,6 @@
+import os
 import pathlib
 import unittest
-
-from PySide6.QtCore import QSettings, QCoreApplication
 
 from src import settings
 from src.model.main_model import MainModel
@@ -10,17 +9,16 @@ from src.view.login_screen import LoginScreen
 
 class LoginScreenTest(unittest.TestCase):
     def setUp(self) -> None:
-        self.env_settings = QSettings()
-        QCoreApplication.setOrganizationName("MercurySeven")
-        QCoreApplication.setApplicationName("SSD")
-        self.path = str(pathlib.Path().absolute()) + "/tests"
+        self.path = os.path.join(str(pathlib.Path().absolute()), "tests")
         self.path = r'%s' % self.path
         pathlib.Path(self.path).mkdir(parents=True, exist_ok=True)
-        self.env_settings.setValue("sync_path", self.path)
-        settings.file_name = "tests/config.ini"
+        settings.file_name = os.path.join(self.path, "config.ini")
         settings.create_standard_settings()
         self.model = MainModel()
         self.login_test = LoginScreen(self.model.network_model)
+
+    def tearDown(self) -> None:
+        os.remove(settings.file_name)
 
     def test_default(self):
         self.assertEqual(self.login_test.userField.text(), self.login_test.model.get_username())
