+++ tests/view/main_view_test.py
@@ -2,8 +2,6 @@ import os
 import pathlib
 import unittest
 
-from PySide6.QtCore import QSettings, QCoreApplication
-
 from src import settings
 from src.model.main_model import MainModel
 from src.view.main_view import MainWindow, MainWidget
@@ -12,15 +10,12 @@ from src.view.main_view import MainWindow, MainWidget
 class MainWindowTest(unittest.TestCase):
 
     def setUp(self) -> None:
-        self.env_settings = QSettings()
-        QCoreApplication.setOrganizationName("MercurySeven")
-        QCoreApplication.setApplicationName("SSD")
-        self.path = str(pathlib.Path().absolute()) + "/tests"
+        self.path = os.path.join(str(pathlib.Path().absolute()), "tests")
         self.path = r'%s' % self.path
         pathlib.Path(self.path).mkdir(parents=True, exist_ok=True)
-        self.env_settings.setValue("sync_path", self.path)
-        settings.file_name = "tests/config.ini"
+        settings.file_name = os.path.join(self.path, "config.ini")
         settings.create_standard_settings()
+
         self.main_model = MainModel()
         self.main_window = MainWindow(self.main_model)
         self.main_widget_test = self.main_window.main_widget
